requirements.txt
aiohttp>=3.8.1
google-adk>=1.0.0
google-genai>=0.1.0
requests>=2.28.0
beautifulsoup
python-dotenv>=1.0.0
pytest>=8.4.1
pytest-asyncio>=0.20.0
markdown>=3.4.0
pytest-asyncio>=0.20.0